pyproject.toml
[project]
name = "nexshift-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.20.0",
    "google-genai>=1.53.0",
    "ortools>=9.14.6206",
    "pydantic>=2.12.5",
    "typing-extensions>=4.15.0",
]
